requirements.txt
alpaca-py==0.43.2
duckdb==1.5.2
fastapi==0.136.0
lxml==6.1.0
numpy==2.2.6
pandas==3.0.2
pandas_market_calendars==5.3.2
pyarrow==23.0.1
python-dotenv==1.2.2
requests==2.33.1
uvicorn==0.44.0
yfinance==1.3.0
